controller tests relocated and fixed
